cleanup: reorganize directories, add MemorySharp so it builds out of the box, move functional code into separate class
