+++ .gitignore
@@ -45,3 +45,8 @@ $RECYCLE.BIN/
 Network Trash Folder
 Temporary Items
 .apdisk
+
+# Visual Studio Stuff
+.vs/
+obj/
+bin/
