+++ H5Tweak.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "H5Tweak", "H5Tweak\H5Tweak.csproj", "{CF228D25-00F2-43B9-86A8-FF659C2372C3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "H5Tweak", "src\H5Tweak\H5Tweak.csproj", "{CF228D25-00F2-43B9-86A8-FF659C2372C3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
